+++ src/main/java/org/itsallcode/holidays/calculator/logic/variants/EasterBasedHoliday.java
@@ -15,19 +15,20 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.itsallcode.holidays.calculator.logic;
+package org.itsallcode.holidays.calculator.logic.variants;
 
 import java.time.LocalDate;
 
-public class OrthodoxEasterBasedHoliday extends PivotDateBasedHoliday {
+import org.itsallcode.holidays.calculator.logic.Easter;
 
-	public OrthodoxEasterBasedHoliday(String category, String name, int offsetInDays) {
-		super("Orthodox Easter", category, name, offsetInDays);
+public class EasterBasedHoliday extends PivotDateBasedHoliday {
+
+	public EasterBasedHoliday(String category, String name, int offsetInDays) {
+		super("Easter", category, name, offsetInDays);
 	}
 
 	@Override
 	public LocalDate of(int year) {
-		return Easter.orthodox(year).plusDays(offsetInDays);
+		return Easter.gauss(year).plusDays(offsetInDays);
 	}
-
 }