+++ src/test/java/org/itsallcode/holidays/calculator/logic/HolidayParserTest.java
@@ -22,11 +22,18 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.time.DayOfWeek;
+import java.time.MonthDay;
 
-import org.itsallcode.holidays.calculator.logic.FloatingHoliday.Direction;
 import org.itsallcode.holidays.calculator.logic.parser.AbbreviationParser.AmbigueAbbreviationException;
 import org.itsallcode.holidays.calculator.logic.parser.AbbreviationParser.InvalidAbbreviationException;
 import org.itsallcode.holidays.calculator.logic.parser.HolidayParser;
+import org.itsallcode.holidays.calculator.logic.variants.EasterBasedHoliday;
+import org.itsallcode.holidays.calculator.logic.variants.FixedDateHoliday;
+import org.itsallcode.holidays.calculator.logic.variants.FloatingHoliday;
+import org.itsallcode.holidays.calculator.logic.variants.FloatingHoliday.Day;
+import org.itsallcode.holidays.calculator.logic.variants.FloatingHoliday.Direction;
+import org.itsallcode.holidays.calculator.logic.variants.Holiday;
+import org.itsallcode.holidays.calculator.logic.variants.OrthodoxEasterBasedHoliday;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -42,28 +49,32 @@ class HolidayParserTest {
 	@Test
 	void category() {
 		assertThat(holidayParser.parse("birthday fixed 7 31 My Birthday"))
-				.isEqualTo(new FixedDateHoliday("birthday", "My Birthday", 7, 31));
+				.isEqualTo(new FixedDateHoliday("birthday", "My Birthday", MonthDay.of(7, 31)));
 		assertThat(holidayParser.parse("birthday fixed 7 31 My Birthday"))
-				.isNotEqualTo(new FixedDateHoliday("holiday", "My Birthday", 7, 31));
+				.isNotEqualTo(new FixedDateHoliday("holiday", "My Birthday", MonthDay.of(7, 31)));
 	}
 
 	@Test
 	void dayOfWeek() {
 		// two letters
 		assertThat(holidayParser.parse("holiday float 4 SU before 12 24 1. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 		// full
 		assertThat(holidayParser.parse("holiday float 4 SUNDAY before 12 24 1. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
-		// lowercase
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
+		// lower case
 		assertThat(holidayParser.parse("holiday float 4 sunday before 12 24 1. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 	}
 
 	@Test
 	void whitespace() {
 		assertThat(holidayParser.parse("   holiday    float  4 SUN\tbefore \t 12 24 1. Advent     "))
-				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 	}
 
 	@Test
@@ -94,14 +105,14 @@ class HolidayParserTest {
 	@Test
 	void nonAmbigueDayOfWeekAbbreviation() {
 		assertThat(holidayParser.parse("holiday float 1 M after 1 1 Non-ambigue M-day"))
-				.isEqualTo(new FloatingHoliday("holiday", "Non-ambigue M-day", 1, DayOfWeek.MONDAY, Direction.AFTER, 1,
-						1));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "Non-ambigue M-day", 1, DayOfWeek.MONDAY, Direction.AFTER, MonthDay.of(1, 1)));
 		assertThat(holidayParser.parse("holiday float 1 W after 1 1 Non-ambigue W-day"))
-				.isEqualTo(new FloatingHoliday("holiday", "Non-ambigue W-day", 1, DayOfWeek.WEDNESDAY, Direction.AFTER,
-						1, 1));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "Non-ambigue W-day", 1, DayOfWeek.WEDNESDAY, Direction.AFTER, MonthDay.of(1, 1)));
 		assertThat(holidayParser.parse("holiday float 1 W after 1 1 Non-ambigue F-day"))
-				.isEqualTo(new FloatingHoliday("holiday", "Non-ambigue F-day", 1, DayOfWeek.WEDNESDAY, Direction.AFTER,
-						1, 1));
+				.isEqualTo(new FloatingHoliday(
+						"holiday", "Non-ambigue F-day", 1, DayOfWeek.WEDNESDAY, Direction.AFTER, MonthDay.of(1, 1)));
 	}
 
 	@Test
@@ -113,54 +124,54 @@ class HolidayParserTest {
 	void notEqualFixed() {
 		// day
 		assertThat(holidayParser.parse("holiday fixed 1 1 Neujahr"))
-				.isNotEqualTo(new FixedDateHoliday("holiday", "Neujahr", 1, 2));
+				.isNotEqualTo(new FixedDateHoliday("holiday", "Neujahr", MonthDay.of(1, 2)));
 		// name
 		assertThat(holidayParser.parse("holiday fixed 1 1 Neujahr"))
-				.isNotEqualTo(new FixedDateHoliday("holiday", "xxx", 1, 1));
+				.isNotEqualTo(new FixedDateHoliday("holiday", "xxx", MonthDay.of(1, 1)));
 	}
 
 	@Test
 	void notEqualFLoating() {
 		// name
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isNotEqualTo(new FloatingHoliday("holiday", "1. Advent AAA", 4, DayOfWeek.SUNDAY, Direction.BEFORE,
-						12, 24));
+				.isNotEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent AAA", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 		// month
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isNotEqualTo(
-						new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 11, 24));
+				.isNotEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(11, 24)));
 		// day
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isNotEqualTo(
-						new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 23));
+				.isNotEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 23)));
 		// day of week
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isNotEqualTo(
-						new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.MONDAY, Direction.BEFORE, 12, 24));
+				.isNotEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 4, DayOfWeek.MONDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 		// offset
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isNotEqualTo(
-						new FloatingHoliday("holiday", "1. Advent", 2, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isNotEqualTo(new FloatingHoliday(
+						"holiday", "1. Advent", 2, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 	}
 
 	@Test
 	void leadingZeros() {
 		assertThat(holidayParser.parse("holiday fixed 01 01 Neujahr"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", 1, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", MonthDay.of(1, 1)));
 		assertThat(holidayParser.parse("holiday float 1 MON before 01 01 Fictional New Year's Monday"))
-				.isEqualTo(new FloatingHoliday("holiday", "Fictional New Year's Monday", 1, DayOfWeek.MONDAY,
-						Direction.BEFORE, 1, 1));
+				.isEqualTo(new FloatingHoliday("holiday", "Fictional New Year's Monday",
+						1, DayOfWeek.MONDAY, Direction.BEFORE, MonthDay.of(1, 1)));
 	}
 
 	@Test
 	void monthNames() {
 		assertThat(holidayParser.parse("holiday fixed Jan 1 Neujahr"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", 1, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", MonthDay.of(1, 1)));
 		assertThat(holidayParser.parse("holiday fixed D 24 XMas Eve"))
-				.isEqualTo(new FixedDateHoliday("holiday", "XMas Eve", 12, 24));
+				.isEqualTo(new FixedDateHoliday("holiday", "XMas Eve", MonthDay.of(12, 24)));
 		assertThat(holidayParser.parse("holiday float 4 SUN before Decem 24 1. Advent"))
 				.isEqualTo(new FloatingHoliday(
-						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
 	}
 
 	@Test
@@ -180,53 +191,93 @@ class HolidayParserTest {
 	@Test
 	void fixedDate() {
 		assertThat(holidayParser.parse("holiday fixed 1 1 Neujahr"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", 1, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", MonthDay.of(1, 1)));
 	}
 
 	@Test
 	void floatingDate() {
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
 				.isEqualTo(new FloatingHoliday(
-						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+						"holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, MonthDay.of(12, 24)));
+
+		assertThat(holidayParser.parse("anniversary float 1 SUN before 12 last-day Last December Sunday"))
+				.isEqualTo(new FloatingHoliday(
+						"anniversary", "Last December Sunday", 1, DayOfWeek.SUNDAY, Direction.BEFORE, 12, Day.LAST));
 	}
 
 	@Test
 	void easter() {
 		assertThat(holidayParser.parse("holiday easter +39 Christi Himmelfahrt"))
-				.isEqualTo(
-						new EasterBasedHoliday("holiday", "Christi Himmelfahrt", +39));
+				.isEqualTo(new EasterBasedHoliday("holiday", "Christi Himmelfahrt", +39));
 	}
 
 	@Test
 	void orthodoxEaster() {
 		assertThat(holidayParser.parse("holiday orthodox-easter -2 Orthodox Good Friday"))
-				.isEqualTo(
-						new OrthodoxEasterBasedHoliday("holiday", "Orthodox Good Friday", -2));
+				.isEqualTo(new OrthodoxEasterBasedHoliday("holiday", "Orthodox Good Friday", -2));
+	}
+
+	@Test
+	void conditionalHoliday() {
+		assertThat(holidayParser.parse(
+				"holiday if DEC 25 is Sat,Sun then fixed DEC 27 Bank Holiday 1"))
+						.isEqualTo(HolidayCalculationTest.BANK_HOLIDAY_DEC_27);
+	}
+
+	@Test
+	void conditionalHolidayWithNegatedDaysOfWeek() {
+		assertThat(holidayParser.parse(
+				"holiday if DEC 25 is not Fri,Sat then fixed DEC 26 Boxing day is extra day off"))
+						.isEqualTo(HolidayCalculationTest.CONDITIONAL_HOLIDAY_WITH_NEGATED_DAYS_OF_WEEK);
+	}
+
+	@Test
+	void alternativeDateHoliday() {
+		final Holiday actual = holidayParser.parse(
+				"holiday either 4 27 or if SUN then fixed 4 26 Koningsdag");
+		assertThat(actual).isEqualTo(HolidayCalculationTest.KONINGSDAG);
+	}
+
+	@Test
+	void alternativeDateHolidayWithNegatedDaysOfWeek() {
+		final Holiday actual = holidayParser.parse(
+				"holiday either 4 27 or if not Mon,Tue,We,Thu,Fri,Sat then fixed 4 26 Koningsdag");
+		assertThat(actual).isEqualTo(HolidayCalculationTest.KONINGSDAG_WITH_NEGATED_DAYS_OF_WEEK);
+	}
+
+	@Test
+	void midsommarAfton_floatingHolidayWithOffsetInDays() {
+		assertThat(holidayParser.parse("holiday float 1 day before 1 Sat before JUN 26 Midsommarafton"))
+				.isEqualTo(HolidayCalculationTest.MIDSOMMARAFTON);
 	}
 
 	@Test
 	void successfulParsing() {
 		assertThat(holidayParser.parse("holiday fixed 1 1 Neujahr"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", 1, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "Neujahr", MonthDay.of(1, 1)));
 		assertThat(holidayParser.parse("holiday fixed 1 6 Heilige Drei Könige"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Heilige Drei Könige", 1, 6));
+				.isEqualTo(new FixedDateHoliday("holiday", "Heilige Drei Könige", MonthDay.of(1, 6)));
 		assertThat(holidayParser.parse("holiday fixed 5 1 1. Mai"))
-				.isEqualTo(new FixedDateHoliday("holiday", "1. Mai", 5, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "1. Mai", MonthDay.of(5, 1)));
 		assertThat(holidayParser.parse("holiday fixed 10 3 Tag der Deutschen Einheit"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Tag der Deutschen Einheit", 10, 3));
+				.isEqualTo(new FixedDateHoliday("holiday", "Tag der Deutschen Einheit", MonthDay.of(10, 3)));
 
 		assertThat(holidayParser.parse("holiday float 4 SUN before 12 24 1. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday("holiday", "1. Advent", 4, DayOfWeek.SUNDAY, Direction.BEFORE,
+						MonthDay.of(12, 24)));
 		assertThat(holidayParser.parse("holiday float 3 SUN before 12 24 2. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "2. Advent", 3, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday("holiday", "2. Advent", 3, DayOfWeek.SUNDAY, Direction.BEFORE,
+						MonthDay.of(12, 24)));
 		assertThat(holidayParser.parse("holiday float 2 SUN before 12 24 3. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "3. Advent", 2, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday("holiday", "3. Advent", 2, DayOfWeek.SUNDAY, Direction.BEFORE,
+						MonthDay.of(12, 24)));
 		assertThat(holidayParser.parse("holiday float 1 SUN before 12 24 4. Advent"))
-				.isEqualTo(new FloatingHoliday("holiday", "4. Advent", 1, DayOfWeek.SUNDAY, Direction.BEFORE, 12, 24));
+				.isEqualTo(new FloatingHoliday("holiday", "4. Advent", 1, DayOfWeek.SUNDAY, Direction.BEFORE,
+						MonthDay.of(12, 24)));
 		assertThat(holidayParser.parse("holiday fixed 12 25 1. Weihnachtstag"))
-				.isEqualTo(new FixedDateHoliday("holiday", "1. Weihnachtstag", 12, 25));
+				.isEqualTo(new FixedDateHoliday("holiday", "1. Weihnachtstag", MonthDay.of(12, 25)));
 		assertThat(holidayParser.parse("holiday fixed 12 26 2. Weihnachtstag"))
-				.isEqualTo(new FixedDateHoliday("holiday", "2. Weihnachtstag", 12, 26));
+				.isEqualTo(new FixedDateHoliday("holiday", "2. Weihnachtstag", MonthDay.of(12, 26)));
 
 		assertThat(holidayParser.parse("holiday easter -48 Rosenmontag"))
 				.isEqualTo(new EasterBasedHoliday("holiday", "Rosenmontag", -48));
@@ -246,12 +297,12 @@ class HolidayParserTest {
 		assertThat(holidayParser.parse("holiday easter +60 Fronleichnam"))
 				.isEqualTo(new EasterBasedHoliday("holiday", "Fronleichnam", +60));
 		assertThat(holidayParser.parse("holiday fixed 8 15 Mariae Himmelfahrt"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Mariae Himmelfahrt", 8, 15));
+				.isEqualTo(new FixedDateHoliday("holiday", "Mariae Himmelfahrt", MonthDay.of(8, 15)));
 		assertThat(holidayParser.parse("holiday fixed 11 1 Allerheiligen"))
-				.isEqualTo(new FixedDateHoliday("holiday", "Allerheiligen", 11, 1));
+				.isEqualTo(new FixedDateHoliday("holiday", "Allerheiligen", MonthDay.of(11, 1)));
 		assertThat(holidayParser.parse("holiday float 1 SUN after 11 20 Totensonntag"))
 				.isEqualTo(new FloatingHoliday(
-						"holiday", "Totensonntag", 1, DayOfWeek.SUNDAY, Direction.AFTER, 11, 20));
+						"holiday", "Totensonntag", 1, DayOfWeek.SUNDAY, Direction.AFTER, MonthDay.of(11, 20)));
 	}
 
 }